deleted files with hometask made via commandLine tool
