--- 2l_NosovAndrei.xcodeproj/xcuserdata/andreinosov.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "7BF7740C-5D54-4A11-B27F-2CA064B581BE"
-   type = "1"
-   version = "2.0">
-</Bucket>
